requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
awsebcli==3.20.5
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
botocore==1.29.81
cached-property==1.5.2
cachetools==5.3.0
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
comm==0.1.3
cryptography==40.0.2
dataclasses-json==0.5.12
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.100.0
fastapi-utils==0.2.1
fastjsonschema==2.17.1
filelock==3.11.0
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.3.3
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
GDAL==3.6.4
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
h5py==3.8.0
hpack==4.0.0
httpcore==0.17.3
httptools==0.6.0
huggingface-hub==0.16.4
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==3.2.0
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.12.0
langchain==0.0.235
langsmith==0.0.7
libclang==16.0.0
loguru==0.7.0
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib-inline==0.1.6
mistune==3.0.1
mmh3==3.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.7
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
paramiko==3.1.0
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
pinecone-text==0.5.2
platformdirs==3.8.1
priority==2.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==5.4.1
pyzmq==25.1.0
quart-cors==0.6.0
regex==2023.6.3
requests==2.26.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.8.5
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
stack-data==0.6.2
starlette==0.27.0
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
termcolor==1.1.0
terminado==0.17.1
texttable==1.6.7
tf-nightly==2.13.0.dev20230413
tf-nightly-macos==2.13.0.dev20230413
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
trio==0.22.2
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.23.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
websockets==11.0.3
Werkzeug==2.3.6
wget==3.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
